dm: core: Use uintptr_t for data in struct udevice_id
Use uintptr_t for data in struct udevice_id as it may use to store pointer
on architecture like CHERI which pointer is not unsigned long integer.

+++ drivers/core/device.c
@@ -1079,7 +1079,7 @@ struct udevice *dev_get_parent(const struct udevice *child)
 	return child->parent;
 }
 
-ulong dev_get_driver_data(const struct udevice *dev)
+uintptr_t dev_get_driver_data(const struct udevice *dev)
 {
 	return dev->driver_data;
 }
