+++ drivers/mfd/atmel-smc.c
@@ -328,9 +328,9 @@ static const struct atmel_hsmc_reg_layout sama5d2_reg_layout = {
 };
 
 static const struct udevice_id atmel_smc_ids[] = {
-	{ .compatible = "atmel,at91sam9260-smc", .data = (ulong)0 },
-	{ .compatible = "atmel,sama5d3-smc", .data = (ulong)&sama5d3_reg_layout },
-	{ .compatible = "atmel,sama5d2-smc", .data = (ulong)&sama5d2_reg_layout },
+	{ .compatible = "atmel,at91sam9260-smc", .data = (uintptr_t)0 },
+	{ .compatible = "atmel,sama5d3-smc", .data = (uintptr_t)&sama5d3_reg_layout },
+	{ .compatible = "atmel,sama5d2-smc", .data = (uintptr_t)&sama5d2_reg_layout },
 	{ /* sentinel */ },
 };
