+++ drivers/pinctrl/exynos/pinctrl-exynos78x0.c
@@ -161,9 +161,9 @@ const struct samsung_pin_ctrl exynos7870_pin_ctrl[] = {
 
 static const struct udevice_id exynos78x0_pinctrl_ids[] = {
 	{ .compatible = "samsung,exynos78x0-pinctrl",
-		.data = (ulong)exynos78x0_pin_ctrl },
+		.data = (uintptr_t)exynos78x0_pin_ctrl },
 	{ .compatible = "samsung,exynos7870-pinctrl",
-		.data = (ulong)exynos7870_pin_ctrl },
+		.data = (uintptr_t)exynos7870_pin_ctrl },
 	{ }
 };
