+++ drivers/pinctrl/nxp/pinctrl-imx93.c
@@ -13,8 +13,8 @@ static struct imx_pinctrl_soc_info imx93_pinctrl_soc_info __section(".data") = {
 };
 
 static const struct udevice_id imx93_pinctrl_match[] = {
-	{ .compatible = "fsl,imx93-iomuxc", .data = (ulong)&imx93_pinctrl_soc_info },
-	{ .compatible = "fsl,imx91-iomuxc", .data = (ulong)&imx93_pinctrl_soc_info },
+	{ .compatible = "fsl,imx93-iomuxc", .data = (uintptr_t)&imx93_pinctrl_soc_info },
+	{ .compatible = "fsl,imx91-iomuxc", .data = (uintptr_t)&imx93_pinctrl_soc_info },
 	{ /* sentinel */ }
 };
