+++ drivers/soc/samsung/exynos-pmu.c
@@ -90,7 +90,7 @@ static const struct udevice_id exynos_pmu_ids[] = {
 	},
 	{
 		.compatible = "samsung,exynos850-pmu",
-		.data = (ulong)&exynos850_pmu_data
+		.data = (uintptr_t)&exynos850_pmu_data
 	},
 	{ /* sentinel */ }
 };
